# backend/src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Bitácora SOC API
  description: API completa para el sistema de Bitácora SOC con gestión de entradas, checklist de turno, usuarios y reportes
  version: 1.0.0
  contact:
    name: Bitácora SOC Team
servers:
  - url: http://192.168.100.50:3000/api
    description: Servidor de desarrollo (IP)
  - url: http://localhost:3000/api
    description: Servidor local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /notes/admin:
    get:
      tags:
        - Notas
      summary: Obtener nota del administrador
      description: Obtiene la nota global del administrador (visible para todos)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Nota del administrador
    
    put:
      tags:
        - Notas
      summary: Actualizar nota del administrador
      description: Actualiza la nota global (solo admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
      responses:
        '200':
          description: Nota actualizada
  /notes/personal:
    get:
      tags:
        - Notas
      summary: Obtener nota personal
      description: Obtiene la nota personal del usuario (privada)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Nota personal
    
    put:
      tags:
        - Notas
      summary: Actualizar nota personal
      description: Actualiza la nota personal del usuario
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
      responses:
        '200':
          description: Nota actualizada
...
  /reports/overview:
    get:
      tags:
        - Reportes
      summary: Vista general de KPIs
      description: Obtiene estadísticas generales del sistema (solo admin)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: days
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: KPIs del sistema
          content:
            application/json:
              schema:
                type: object
                properties:
                  period:
                    type: string
                  entriesByType:
                    type: object
                  incidentsByUser:
                    type: array
                  topTags:
                    type: array
                  redsByService:
                    type: array
                  entriesTrend:
                    type: array
                  totalUsers:
                    type: integer
                  totalChecks:
                    type: integer
  /reports/export-entries:
    get:
      tags:
        - Reportes
      summary: Exportar entradas a CSV
      description: Exporta entradas filtradas a formato CSV (solo admin)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Archivo CSV
          content:
            text/csv:
              schema:
                type: string
  /config:
    get:
      tags:
        - Configuración
      summary: Obtener configuración
      description: Obtiene la configuración global de la aplicación
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Configuración actual
    
    put:
      tags:
        - Configuración
      summary: Actualizar configuración
      description: Actualiza la configuración global (solo admin)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                guestModeEnabled:
                  type: boolean
                guestMaxDurationDays:
                  type: integer
                  minimum: 1
                  maximum: 30
                shiftCheckCooldownHours:
                  type: integer
                  minimum: 1
                  maximum: 24
                logoUrl:
                  type: string
      responses:
        '200':
          description: Configuración actualizada
...